requirements.txt
ag-ui-protocol==0.1.10
annotated-types==0.7.0
anthropic==0.74.0
anyio==4.11.0
argcomplete==3.6.3
attrs==25.4.0
authlib==1.6.5
backports-tarfile==1.2.0
beartype==0.22.5
boto3==1.41.0
botocore==1.41.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.4.1
click==8.3.1
cohere==5.20.0
colorama==0.4.6
cryptography==46.0.3
cyclopts==4.2.4
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
docutils==0.22.3
email-validator==2.3.0
eval-type-backport==0.3.0
exceptiongroup==1.3.0
executing==2.2.1
fastavro==1.12.1
fastmcp==2.13.1
filelock==3.20.0
fsspec==2025.10.0
genai-prices==0.0.42
google-auth==2.43.0
google-genai==1.51.0
googleapis-common-protos==1.72.0
griffe==1.15.0
groq==0.35.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==1.1.4
idna==3.11
importlib-metadata==8.7.0
invoke==2.2.1
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.3.0
jiter==0.12.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.7.0
logfire==4.15.0
logfire-api==4.15.0
markdown-it-py==4.0.0
mcp==1.21.2
mdurl==0.1.2
mistralai==1.9.11
more-itertools==10.8.0
nexus-rpc==1.1.0
ollama==0.6.1
openai==2.8.1
openapi-pydantic==0.5.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-httpx==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
packaging==25.0
pathable==0.4.4
pathvalidate==3.3.1
platformdirs==4.5.0
prompt-toolkit==3.0.52
protobuf==6.33.1
py-key-value-aio==0.2.8
py-key-value-shared==0.2.8
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-ai==1.20.0
pydantic-ai-slim==1.20.0
pydantic-core==2.41.5
pydantic-evals==1.20.0
pydantic-graph==1.20.0
pydantic-settings==2.12.0
pygments==2.19.2
pyjwt==2.10.1
pypdf==6.6.0
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pyyaml==6.0.3
referencing==0.36.2
requests==2.32.5
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.29.0
rsa==4.9.1
s3transfer==0.14.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.3
starlette==0.50.0
temporalio==1.19.0
tenacity==9.1.2
tokenizers==0.22.1
tqdm==4.67.1
typer-slim==0.20.0
types-protobuf==6.32.1.20251105
types-requests==2.32.4.20250913
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
websockets==15.0.1
wrapt==1.17.3
zipp==3.23.0
